+++ lib/store_model/combine_errors_strategies/mark_invalid_error_strategy.rb
@@ -10,8 +10,8 @@ module StoreModel
       # @param base_errors [ActiveModel::Errors] errors object of the parent record
       # @param _store_model_errors [ActiveModel::Errors] errors object of the
       # StoreModel::Model attribute
-      def call(attribute, base_errors, _store_model_errors)
-        base_errors.add(attribute, :invalid)
+      def call(attribute, base_errors, store_model_errors)
+        base_errors.add(attribute, :invalid, errors: store_model_errors)
       end
     end
   end